Rendering Improvements
Most of this update is just to vastly improve the slow rendering we had. 
To do this we added instance rendering on everything we render, perhaps there will an option in the far far future, if we really dont want instanced rendering, we can set and save a batch size.
 The other thing was to speed up the module retrieval. This had gone through a couple of revisions but the current version is only updates what modules have been added/removed instead of starting from scratch every time we anted to update. The option to force an update from scratch will be coming soon