Refactoring code
Added interfaces which interface with the a module
Renamed modules, interfaces, resources and systems to be 1 letter shorthands so the code isnt so verbose
Added a pointer to each module the entity uses in the entity header but isnt really used
Changed the modules to use shared pointers instead of unique pointer and raw pointers so I can streamline returns and stuff. This may potentially be bad when removing stuff because I don't know where the dangling references are. There are still some raw pointers so I know when they get removed but I keep working on this